+++ xds/src/main/java/io/grpc/xds/internal/certprovider/CertProviderSslContextProvider.java
@@ -22,8 +22,8 @@ import io.envoyproxy.envoy.extensions.transport_sockets.tls.v3.CommonTlsContext;
 import io.envoyproxy.envoy.extensions.transport_sockets.tls.v3.CommonTlsContext.CertificateProviderInstance;
 import io.grpc.xds.Bootstrapper.CertificateProviderInfo;
 import io.grpc.xds.EnvoyServerProtoData.BaseTlsContext;
-import io.grpc.xds.internal.sds.CommonTlsContextUtil;
-import io.grpc.xds.internal.sds.DynamicSslContextProvider;
+import io.grpc.xds.internal.security.CommonTlsContextUtil;
+import io.grpc.xds.internal.security.DynamicSslContextProvider;
 import java.security.PrivateKey;
 import java.security.cert.X509Certificate;
 import java.util.List;
