+++ xds/src/main/java/io/grpc/xds/internal/security/Closeable.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.xds.internal.sds;
+package io.grpc.xds.internal.security;
 
 public interface Closeable extends java.io.Closeable {
